pyproject.toml
[project]
authors = [{name = "jlnav", email = "jnavarro@anl.gov"}]
description = "Add a short description here"
name = "pdf2json"
requires-python = ">= 3.10"
version = "0.1.0"
dependencies = ["transformers", "pillow", "pymupdf"]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-arm64", "linux-64"]

[tool.pixi.pypi-dependencies]
pdf2json = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.environments]
default = []
dev = ["dev"]

[tool.pixi.dependencies]
python = ">=3.10.0"
pytorch = ">=2.5.1,<3"
lightning = ">=2.5.0.post0,<3"
torchvision = ">=0.20.1,<0.21"
timm = ">=1.0.14,<2"
nltk = ">=3.9.1,<4"
python-levenshtein = ">=0.26.1,<0.27"

[tool.pixi.feature.dev.dependencies]
ipdb = ">=0.13.13,<0.14"
pre-commit = ">=4.0.1,<5"
pip = ">=24.3.1,<25"
